Add files via upload
This data set that includes data on conferences, the number of documents for each conference and their citations, and the rating of conferences. The first tab of the file contains data on conferences, the number of documents and the number of citations. The second tab contains the same data for China, the third for the United States, and the fourth tab contains the division of 249 conferences into 4 quartiles based on the CPP indicator (citation per paper).